refactor(config): move up config
